requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
redis>=5.0.1
sentence-transformers>=2.3.1
scikit-learn>=1.4.0
numpy>=1.26.3
httpx>=0.26.0
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
requests>=2.31.0
chromadb>=0.4.22
PyYAML>=6.0
